rustgen: scaffolding for boolean and option
Extends implementation in tcc/rust/kernel library to facilitate
future handling of boolean value parsing, as well as laying the
groundwork required for automatically derived Option<T> codec
implementations.

Enriches and refactors the ParseError enumerator to allow for
cleaner boilerplate at callsite, and to allow extensibility in
future revisions of the model.