rustgen: revise model for hexstrings
Refactor `hex!` macro defined in rust_runtime/src/parse.rs
to use the locally-scoped HexString type, as the
previous definition only worked when the macro was used in the runtime
crate, and not in the generated code.
Note: this redefinition requires that HexString is
imported, and for this reason the definition
is exposed under the toplevel crate namespace for rust_runtime
through a `pub use` in the `lib.rs` for rust_runtime

Adds more (derived) traits for HexString to permit its use in
certain coding patterns it was previously inapplicable for
(specificically, `assert_eq!` checks in test code)

Refines model of generated-rust preludes to allow for `#[mod_use]`
annotations to be included in codec modules.

Adds `_HexString` shorthand in `Ast.Type` to match `_String` and `_str`
and defines HexString as a pseudo-native `str_kind` without distinction.

Refactors `mk_impl_encode` in `Generator.Gen` to return HexStrings
rather than Strings, as a portable encoding
for non-UTF8 binary blobs representing various
primitive types has not yet been fully worked out.
Namely, binary blobs that are invalid unicode strings are rejected by
the rust compiler in the present runtime implementation,
so using HexStrings as a stopgap is the most effective way of
solving this problem in the short-term.

Adds a simple Makefile to `codec_generator/` with rules for
generating rust codec modules, and for producing the intermediate
compilation objects for ocamllsp/merlin to do its job properly when the
compiler code is changed.

Adds updated (albeit not 100% working) versions of the generated codec
`.rs` files. Note that the `main.rs` file is not currently working.