rustgen: Refactor parser offset tracking
Adds new module to rust_runtime, `internal.rs`, which defines
structures and traits that are used to track
the monotonically increasing offset of a
parser with an immutable buffer,
through an internally mutable container around a `usize`
More specifiically, the Offset struct and Marker trait
are the main exposed definitions of this module.

Refactors ByteParser to be parametric in its offset-counter type,
which is required to be an implementor
of the Marker trait (from `internal.rs`) and
defaults to `Offset` (ibid) when not
given explicitly. The ToParser trait
is refactored similarly.

Refactors logic in ByteParser methods to use the new
model for offset-tracking and increments.

Cleans up preambles in `main.rs` and `lib.rs`