rustgen: add f64 and RangedFloat
Refactors OutOfRange to operate generically
over the widest version of its restriction type,
either i64 or f64 (but not type-level constrained to be either)

Refactors parse::errors::ExternalErrorKind to include
int- and float-specialized RangeViolations

Moves OutOfRange to bounds.rs

Implements float.rs, with f64 and RangedFloat Encode/Decode

NOTE: RangedFloat bounds are u64 bit-reps of the equivalent f64
values, which cannot be avoided. This is unfortunate but a better
design is not obvious at this time.

Adds FixedLength, Encode and Decode for u64 so that
f64 can use such definitions as a direct proxy, even if
not used directly in the schema.