rustgen: gen macros for ADTs, extend tag parsing
Reframes approach for handling codec generation of discriminated unions
(enums) as one of generation rather
than derivation:

Adds `rust_runtime::adt` module with two
macros, `cstyle!` and `data!`, which
generate both the definitions for,
and the required trait implementations
we would otherwise have to
`#[derive(...)]`, respectively
on data-free (nullary or C-style)
and data-full (non-nullary or Data)
enum types.
TODO: clean up the syntax and review the approach

Removes scaffolding for derive_decode handling of enums.

Extends the capabilities of tagword parsing to cover 1-, 2-, and 4-byte
discriminants more cleanly

Adds associated function to Dynamic that allows
us to create Dynamic values by hand for testing
purposes

Adds temporary testing crate `sandbox` to allow for testing the infrastructure
of the rust_runtime against
hand-written codecs before the
syntax is cemented as the API that
the OCaml rustgen pipeline
has to conform to

Fixes small bug with unit-like enum decode-derive
TODO: make sure this bug does not exist
for the other derive macros