rustgen: documentation, copy benchmark
Adds more detailed documentation for
the library crate `runtime` itself, as well
as to builder::lazy and conv.

Adds a benchmark for the comparitive
performance of std::io::copy vs
std::io::Write::write_all when copying
from a &Vec<u8> to a generic writer.

Adds a `write_any` method for `Encode` that takes a `W: std::io::Write`
generic parameter and is analogous
to `write` over `&mut W`.

Adds a warning message to the default
implementation of EncodeLength::enc_len
to indicate it has not been overridden and is
therefore inefficient

Adds `to_bytes_full` method to EncodeLength as a
refinement of `Encode::to_bytes`, which
initializes the capacity of the newly created
Vec<u8> buffer to the value returned by
`enc_len`; this is only efficient if
`enc_len` is overriden and is otherwise
unjustifiably inefficient.

TODO: consider revising strategy for Builder-based encode using
std::io or otherwise commonplace traits
such as Write.