rustgen: data macro unit fix
Introduces a syntactic workaround to
permit unit-struct payloads
in data! macro invocations
by introducing disambiguator character `_` between
variant name and variant-separating comma.

This is implemented both in the runtime and ast layer in the pp logic
for struct definitions.

This is technically optional if used for the final variant,
but there is currently no logic
in the AST pp code to detect
and handle this case differently.

Unifies non-record cases in fuse_record

TODO: look into a redefinition of data! using
more advanced macro techniques to avoid
having to alter the syntax to solve the ambiguity

TODO: consider a different marker than `_` for solving the ambiguous
comma problem