rust-gen: data macro patch, len rename
Refactors `data!` adt-macro to allow for
unit structs without special syntax.

NOTE: as a result of this change, the trailing comma
for the final variant in a data! macro block
is now mandatory

Renames `len()` method of `Estimable` to `estimate()`
in order to avoid lexical ambiguity
with type-associated `.len()` functions.