rustgen: explicit submodule name for data
Refactors `data!` macro so that the name of the
payload sub-module is a mandatory parameter,
which allows for multiple `data!`
invocations in a single module.

Refactors generator and API to introduce
a new string field to `UnData` AST nodes
for the name of the payload sub-module.

Refactors Generator.Ctxt to have a single,
unified State module for all contextual
metadata, embedding modules
Ctxt.S and Ctxt.T under Ctxt.State

Redefines `Ctxt.State.T.union` so that it is non-weak polymorphic.

Reformats compiler source
Regenerates test-codecs