rustgen: Redesign of Encode
Redesigns the core API of the Encode trait
so that the primitive required method is
`write_to`, which accepts a generic type as
the destination of serialization, bound by
the newly-introduced `Target` trait, and
with a return value indicating the number of
bytes written.

Introduces Target trait, which is an analogue of std::io::Write,
as an abstraction layer to allow for later
revisions that would be more complex
if std::io::Write were used instead
as the trait bound for the Encode::write_to
method destination.

Adds Target as a dependent bound for implementing Builder, adds
Target implementations for
MemoBuilder and StrictBuilder

Re-implements MemoBuilder using a single-buffer approach, with potential
to adapt a more hierarchical model
of segmentation.

Introduces an internal collection type SplitVec for use
in the redefinition of MemoBuilder.
The model of a SplitVec is an append-only buffer that
is virtually divided into fixed-length sub-segments.
This division is linear rather than hierarchical,
but a rope-like model is theoretically
possible as a redesign.
TODO: reconcile duplication in MemoBuffer model

Prunes unused or inefficient modules and traits:
  - LazyBuilder
  - TransientBuilder
  - rug-based N/Z
  - OwnedBuilder

Removes extraneous benches:
  - zarith_bench
  - builder_bench

Propogates necessary changes in design to
the AST and generator for rust.

Adds Snippets module to reduce open-coding of
generative logic that depends on the runtime API
more strongly than Rustlang proper,
so that future changes can be made more
cleanly.
TODO: continue reduction of open-coded dependent logic

Re-generates all sandbox and generated_code_to_remove
modules according to updated compiler logic.

Replaces OwnedBuilder in AST with StrictBuilder

Adds scaffolding type constructor `usize` to
Rust.Ast.Type.t variants

Adds summation expressions to Rust.Ast.Expr.t