tsgen: sequence type, proxy patch
Adds experimental definition for Sequence class with
a Kind-generic for the variety of sequence (fixed, bounded,
unbounded).

Redefines the `Proxy` type to operate on objects with extensible
fields, specified as a default-`{}` second generic argument

Refactors option.ts to conform to new definition of Proxy.

Rustgen:
patches known bug in FixSeq behavior on decode, to be tested
TODO: update README to reflect this patch